refactor: move Driver's methods to actions modules
