+++ handofcats/driver.py
@@ -1,6 +1,11 @@
 import typing as t
 from .injector import Injector
-from .actions import TargetFunction, ArgumentParser
+from .types import (
+    TargetFunction,
+    ArgumentParser,
+    CustomizeSetupFunction,
+    CustomizeActivateFunction,
+)
 from . import customize
 from prestring.naming import titleize
 
@@ -21,102 +26,17 @@ class Driver:
 
         fargs, rest = first_parser.parse_known_args(argv)
         if fargs.expose:  # "--expose" ?
-            return self._run_expose_action(
-                fn, rest, inplace=fargs.inplace, typed=fargs.typed
+            from .actions import codegen
+
+            return codegen.run_as_single_command(
+                fn, rest, inplace=fargs.inplace, typed=fargs.typed,
             )
         else:
-            return self._run_command_line(fn, rest)
+            from .actions import commandline
 
-    def _run_command_line(
-        self, fn: TargetFunction, argv: t.Optional[str] = None
-    ) -> t.Any:
-        from .actions import commandline
-
-        m = commandline.setup_module()
-
-        customizations = []
-        if not self.ignore_logging:
-            # TODO: include generated code, emitted by `--expose`
-            customizations.append(customize.logging_setup)
-
-        parser, activate_functions = self.setup_parser(
-            m, fn, customizations=customizations
-        )
-        args = parser.parse_args(argv)
-        params = vars(args).copy()
-
-        for activate in activate_functions:
-            activate(params)
-        return fn(**params)
-
-    def _run_expose_action(
-        self,
-        fn: TargetFunction,
-        argv: t.Optional[str] = None,
-        *,
-        outname: str = "main",
-        inplace: bool = False,
-        typed: bool = False,
-    ) -> None:
-        """ generate main() code
-
-        something like
-
-        ```
-        def main(argv=None):
-            import argparse
-
-            parser = argparse.ArgumentParser(prog=hello.__name__, description=hello.__doc__)
-            parser.print_usage = parser.print_help
-
-            # adding code, by self.setup_parser(). e.g.
-            # parser.add_argument('--name', required=False, default='world', help="(default: 'world')")
-            # parser.add_argument('--debug', action="store_true")
-
-            args = parser.parse_args(argv)
-            params = vars(args).copy()
-            return hello(**params)
-
-        if __name__ == "__main__":
-            main()
-        ```
-        """
-        from .actions import codegen
-
-        m = codegen.setup_module()
-
-        if fn.__name__ == outname:
-            outname = titleize(outname)  # main -> Main
-
-        if typed:
-            m.sep()
-            m.from_("typing").import_("Optional, List  # noqa: E402")
-            m.sep()
-            mdef = m.def_(
-                outname, "argv: Optional[List[str]] = None", return_type="None"
+            return commandline.run_as_single_command(
+                self.setup_parser, fn, rest, ignore_logging=self.ignore_logging
             )
-        else:
-            mdef = m.def_(outname, "argv=None")
-
-        # def main(argv=None):
-        with mdef:
-            parser, _ = self.setup_parser(m, fn, customizations=[])
-
-            # args = parser.parse_args(argv)
-            args = m.let("args", parser.parse_args(m.symbol("argv")))
-
-            # params = vars(args).copy()
-            _ = m.let("params", m.symbol("vars")(args).copy())
-
-            # return fn(**params)
-            m.return_(f"{fn.__name__}(**params)")
-
-        # if __name__ == "__main__":
-        with m.if_("__name__ == '__main__'"):
-            # main()
-            m.stmt(f"{outname}()")
-
-        codegen.emit(m, fn, inplace=inplace)
 
     def setup_parser(
         self,
@@ -124,8 +44,8 @@ class Driver:
         fn: TargetFunction,
         argv=None,
         *,
-        customizations: t.Optional[t.List[customize.SetupFunction]] = None,
-    ) -> t.Tuple[ArgumentParser, t.List[customize.ActivateFunction]]:
+        customizations: t.Optional[t.List[CustomizeSetupFunction]] = None,
+    ) -> t.Tuple[ArgumentParser, t.List[CustomizeActivateFunction]]:
         # import argparse
         argparse = m.import_("argparse")
         m.sep()
@@ -266,8 +186,8 @@ class MultiDriver:
         m,
         functions: t.List[TargetFunction],
         *,
-        customizations: t.Optional[t.List[customize.SetupFunction]] = None,
-    ) -> t.Tuple[ArgumentParser, t.List[customize.ActivateFunction]]:
+        customizations: t.Optional[t.List[CustomizeSetupFunction]] = None,
+    ) -> t.Tuple[ArgumentParser, t.List[CustomizeActivateFunction]]:
         # import argparse
         argparse = m.import_("argparse")
         m.sep()
